Rename `classpath` -> `build-tools`
This will make it simpler if we add more common stuff to this module
later.
