+++ settings.gradle
@@ -19,17 +19,17 @@ def jdbcModules = [
         'postgresql'
 ]
 
-include 'test-resources-core'
-include 'test-resources-embedded'
-include 'test-resources-client'
 include 'test-resources-bom'
+include 'test-resources-build-tools'
+include 'test-resources-core'
+include 'test-resources-client'
+include 'test-resources-embedded'
 include 'test-resources-hivemq'
 include 'test-resources-kafka'
 include 'test-resources-mongodb'
 include 'test-resources-neo4j'
 include 'test-resources-testcontainers'
 include 'test-resources-server'
-include 'test-resources-classpath'
 
 jdbcModules.each {
     String projectName = "test-resources-jdbc-$it"
