+++ test-resources-build-tools/src/main/java/io/micronaut/testresources/buildtools/MavenDependency.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micronaut.testresources.classpath;
+package io.micronaut.testresources.buildtools;
 
 /**
  * Represents a Maven dependency using its group, artifact
